requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
schedule==1.2.0 